+++ sdk/dotnet/AVS/V20200717Preview/GetWorkloadNetworkDnsService.cs
@@ -80,7 +80,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
         /// <summary>
         /// NSX revision number.
         /// </summary>
-        public readonly int? Revision;
+        public readonly double? Revision;
         /// <summary>
         /// DNS Service status.
         /// </summary>
@@ -108,7 +108,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
 
             string provisioningState,
 
-            int? revision,
+            double? revision,
 
             string status,
