+++ sdk/dotnet/Compute/Latest/GetSnapshot.cs
@@ -50,7 +50,7 @@ namespace Pulumi.AzureNextGen.Compute.Latest
         /// <summary>
         /// The size of the disk in bytes. This field is read only.
         /// </summary>
-        public readonly int DiskSizeBytes;
+        public readonly double DiskSizeBytes;
         /// <summary>
         /// If creationData.createOption is Empty, this field is mandatory and it indicates the size of the disk to create. If this field is present for updates or creation with other options, it indicates a resize. Resizes are only allowed if the disk is not attached to a running VM, and can only increase the disk's size.
         /// </summary>
@@ -130,7 +130,7 @@ namespace Pulumi.AzureNextGen.Compute.Latest
 
             string? diskAccessId,
 
-            int diskSizeBytes,
+            double diskSizeBytes,
 
             int? diskSizeGB,
