+++ sdk/dotnet/Compute/V20180930/GetDisk.cs
@@ -46,7 +46,7 @@ namespace Pulumi.AzureNextGen.Compute.V20180930
         /// <summary>
         /// The number of IOPS allowed for this disk; only settable for UltraSSD disks. One operation can transfer between 4k and 256k bytes.
         /// </summary>
-        public readonly int? DiskIOPSReadWrite;
+        public readonly double? DiskIOPSReadWrite;
         /// <summary>
         /// The bandwidth allowed for this disk; only settable for UltraSSD disks. MBps means millions of bytes per second - MB here uses the ISO notation, of powers of 10.
         /// </summary>
@@ -116,7 +116,7 @@ namespace Pulumi.AzureNextGen.Compute.V20180930
         private GetDiskResult(
             Outputs.CreationDataResponse creationData,
 
-            int? diskIOPSReadWrite,
+            double? diskIOPSReadWrite,
 
             int? diskMBpsReadWrite,
