+++ sdk/dotnet/Compute/V20190301/Disk.cs
@@ -25,7 +25,7 @@ namespace Pulumi.AzureNextGen.Compute.V20190301
         /// The number of IOPS allowed for this disk; only settable for UltraSSD disks. One operation can transfer between 4k and 256k bytes.
         /// </summary>
         [Output("diskIOPSReadWrite")]
-        public Output<int?> DiskIOPSReadWrite { get; private set; } = null!;
+        public Output<double?> DiskIOPSReadWrite { get; private set; } = null!;
 
         /// <summary>
         /// The bandwidth allowed for this disk; only settable for UltraSSD disks. MBps means millions of bytes per second - MB here uses the ISO notation, of powers of 10.
@@ -37,7 +37,7 @@ namespace Pulumi.AzureNextGen.Compute.V20190301
         /// The size of the disk in bytes. This field is read only.
         /// </summary>
         [Output("diskSizeBytes")]
-        public Output<int> DiskSizeBytes { get; private set; } = null!;
+        public Output<double> DiskSizeBytes { get; private set; } = null!;
 
         /// <summary>
         /// If creationData.createOption is Empty, this field is mandatory and it indicates the size of the disk to create. If this field is present for updates or creation with other options, it indicates a resize. Resizes are only allowed if the disk is not attached to a running VM, and can only increase the disk's size.
@@ -197,7 +197,7 @@ namespace Pulumi.AzureNextGen.Compute.V20190301
         /// The number of IOPS allowed for this disk; only settable for UltraSSD disks. One operation can transfer between 4k and 256k bytes.
         /// </summary>
         [Input("diskIOPSReadWrite")]
-        public Input<int>? DiskIOPSReadWrite { get; set; }
+        public Input<double>? DiskIOPSReadWrite { get; set; }
 
         /// <summary>
         /// The bandwidth allowed for this disk; only settable for UltraSSD disks. MBps means millions of bytes per second - MB here uses the ISO notation, of powers of 10.
