+++ sdk/dotnet/ContainerRegistry/Latest/Outputs/TargetResponseResult.cs
@@ -20,7 +20,7 @@ namespace Pulumi.AzureNextGen.ContainerRegistry.Latest.Outputs
         /// <summary>
         /// The number of bytes of the content. Same as Size field.
         /// </summary>
-        public readonly int? Length;
+        public readonly double? Length;
         /// <summary>
         /// The MIME type of the referenced object.
         /// </summary>
@@ -36,7 +36,7 @@ namespace Pulumi.AzureNextGen.ContainerRegistry.Latest.Outputs
         /// <summary>
         /// The number of bytes of the content. Same as Length field.
         /// </summary>
-        public readonly int? Size;
+        public readonly double? Size;
         /// <summary>
         /// The tag name.
         /// </summary>
@@ -54,7 +54,7 @@ namespace Pulumi.AzureNextGen.ContainerRegistry.Latest.Outputs
         private TargetResponseResult(
             string? digest,
 
-            int? length,
+            double? length,
 
             string? mediaType,
 
@@ -62,7 +62,7 @@ namespace Pulumi.AzureNextGen.ContainerRegistry.Latest.Outputs
 
             string? repository,
 
-            int? size,
+            double? size,
 
             string? tag,
